requirements.txt
streamlit>=1.28.0
pyvis>=0.3.2
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pathlib2>=2.3.7
